refactor: 주문 도메인 구조 변경
