+++ build.gradle.kts
@@ -16,6 +16,10 @@ java {
     }
 }
 
+repositories {
+    mavenCentral()
+}
+
 subprojects {
     repositories {
         mavenCentral()
