+++ modules/order/order-api/src/main/kotlin/me/hi0yoo/commerce/order/application/dto/PayOrderCommand.kt
@@ -1,8 +1,6 @@
 package me.hi0yoo.commerce.order.application.dto
 
-import me.hi0yoo.commerce.order.domain.PayMethod
-
 data class PayOrderCommand(
     val orderId: String,
-    val payMethod: PayMethod,
+    val payMethod: String,
 )
