+++ modules/order/order-api/src/main/kotlin/me/hi0yoo/commerce/order/application/service/PaySystemFactory.kt
@@ -1,4 +1,4 @@
-package me.hi0yoo.commerce.order.application
+package me.hi0yoo.commerce.order.application.service
 
 import me.hi0yoo.commerce.order.application.port.out.PaySystemPort
 import me.hi0yoo.commerce.order.domain.PayMethod
